backend/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
